qcacmn: Fix issues in monitor path
1. Replenish path issue:

a. when number of required buffers passed to replenish API is '1',
host does not replenish '1' buffer.

b. If no. of available entries in ring are '0', return sw desc
   back to pool.

2. 32-bit compilation issue

3. Block ppdu desc deliver API when pktlog is enabled for BE

+++ hal/wifi3.0/be/hal_be_api_mon.h
@@ -179,10 +179,10 @@ void hal_mon_buff_addr_info_set(hal_soc_handle_t hal_soc_hdl,
 				void *mon_desc_addr,
 				qdf_dma_addr_t phy_addr)
 {
-	uint32_t paddr_lo = ((u64)phy_addr & 0x00000000ffffffff);
-	uint32_t paddr_hi = ((u64)phy_addr & 0xffffffff00000000) >> 32;
-	uint32_t vaddr_lo = ((u64)mon_desc_addr & 0x00000000ffffffff);
-	uint32_t vaddr_hi = ((u64)mon_desc_addr & 0xffffffff00000000) >> 32;
+	uint32_t paddr_lo = ((uintptr_t)phy_addr & 0x00000000ffffffff);
+	uint32_t paddr_hi = ((uintptr_t)phy_addr & 0xffffffff00000000) >> 32;
+	uint32_t vaddr_lo = ((uintptr_t)mon_desc_addr & 0x00000000ffffffff);
+	uint32_t vaddr_hi = ((uintptr_t)mon_desc_addr & 0xffffffff00000000) >> 32;
 
 	HAL_MON_PADDR_LO_SET(mon_entry, paddr_lo);
 	HAL_MON_PADDR_HI_SET(mon_entry, paddr_hi);
